requirements.txt
flask
oracledb
python-dotenv
twelvelabs
oci
Pillow
geopy
